chore: lint code

+++ src/components/sessionDenyButton.ts
@@ -6,15 +6,22 @@ import {
 } from "discord.js";
 import Sessions from "../models/sessions.js";
 import type { MoulinetteButtonComponent } from "../types/component.js";
+import { replyError } from "../util/functions.js";
 
-export default <MoulinetteButtonComponent>{
+export default {
 	data: new ButtonBuilder()
 		.setCustomId(`session:deny-00000000000`)
 		.setStyle(ButtonStyle.Danger)
 		.setLabel("No"),
-	regexp: /^session:deny-[a-f0-9]{11}$/,
+	pattern: /^session:deny-[a-f0-9]{11}$/,
 	async execute(_client, interaction) {
-		const token = interaction.customId.match(/[a-f0-9]{11}$/)?.[0];
+		const token = /[a-f0-9]{11}$/.exec(interaction.customId)?.[0];
+		
+		if (!token)
+			return interaction.reply(
+				replyError("Invalid session token, please try again later..."),
+			);
+
 		const session = await Sessions.findByIdAndDelete(token);
 
 		if (!session)
@@ -29,7 +36,7 @@ export default <MoulinetteButtonComponent>{
 				flags: MessageFlags.Ephemeral,
 			});
 
-		interaction.update({
+		await interaction.update({
 			embeds: [
 				new EmbedBuilder()
 					.setColor("Red")
@@ -41,4 +48,4 @@ export default <MoulinetteButtonComponent>{
 
 		return;
 	},
-};
+} as MoulinetteButtonComponent;
